tests/clean
